+++ src/engine/Traits/Index.ts
@@ -1,4 +1,4 @@
 export * from './CapturePointer';
-export * from './ChunkSystemTileMapCollisionDetection';
+export * from './CellMapCollisionDetection';
 export * from './OffscreenCulling';
 export * from './TileMapCollisionDetection';